fix(TagManager): separate code snippet extensions to separate file
